Fixes on CSS and break header component
