+++ app/javascript/packs/components/menu-item.vue
